[MIG] web_ir_actions_act_window_message: Migration to 15.0
