--- web_ir_actions_act_window_message/views/templates.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2017 Therp BV, ACSONE SA/NV
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-<odoo>
-    <template
-        id="assets_backend"
-        name="web_ir_actions_act_window_message assets"
-        inherit_id="web.assets_backend"
-    >
-        <xpath expr="." position="inside">
-            <script
-                type="text/javascript"
-                src="/web_ir_actions_act_window_message/static/src/js/web_ir_actions_act_window_message.js"
-            />
-        </xpath>
-    </template>
-</odoo>
